Restructure files
This is inspired by https://towardsdatascience.com/how-to-machine-leaning-pipeline-beginner-2a736595cbfd.
