+++ .gitignore
@@ -1,2 +1,3 @@
 .idea
 .env
+tiny-imagenet-200/
