+++ README.md
@@ -1,0 +1,13 @@
+ML pipeline:
+1. Data extraction
+2. Data pre-processing
+3. EDA & further data processing
+4. Model training
+5. Model evaluation
+
+Instructions to run:
+1. `git clone https://github.com/lonerly666/smartTech_CA1.git <rootDirName>`
+2. `cd <rootDirName>`
+3. Create .env file as follows:
+`ROOT_DIR=${PWD}`
+4. Run the pipeline using `./run.sh`. Please execute this file only from `<rootDirName>`
