+++ run.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+python ./src/run.py
