+++ src/run.py
@@ -1,0 +1,3 @@
+from ml_module.data_extraction import extract_data
+
+(x_train, y_train), (x_val, y_val) = extract_data()
